refact: move ark server functions to utils/
