+++ apps/mobile/hooks/useArkWallet.ts
@@ -1,10 +1,10 @@
 import { useQuery } from '@tanstack/react-query'
 
 import { openArkWallet } from '@/api/ark'
-import { getArkServer } from '@/constants/ark'
 import { ensureArkDatadir } from '@/storage/arkDatadir'
 import { getArkMnemonic } from '@/storage/encrypted'
 import { useArkStore } from '@/store/ark'
+import { getArkServer } from '@/utils/ark'
 
 async function ensureWalletOpen(accountId: string): Promise<true> {
   const { accounts, serverAccessTokens } = useArkStore.getState()
